# ff-split/pkg/api/openapi.yaml
openapi: 3.0.3
info:
  title: FinFlow Split Service API
  description: |
    API для сервиса управления совместными расходами в FinFlow.
    
    Сервис предоставляет функциональность для:
    - Управления мероприятиями и событиями
    - Создания и отслеживания транзакций
    - Разделения расходов между пользователями
    - Управления задачами и активностями
    - Работы с категориями и иконками
    
    ## Аутентификация
    
    Защищенные endpoints требуют аутентификации через Bearer токен:
    ```
    Authorization: Bearer <access_token>
    ```
  version: 1.0.0
  contact:
    name: FinFlow Team
  license:
    name: MIT
servers:
  - url: http://localhost:8081
    description: Development server
...
security:
  - BearerAuth: []
paths:
  /api/v1/event:
    get:
      tags:
        - events
      summary: Получить список мероприятий
      description: Возвращает список всех мероприятий
      operationId: getEvents
      responses:
        '200':
          description: Список мероприятий
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - events
      summary: Создать мероприятие
      description: Создает новое мероприятие
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '201':
          description: Мероприятие создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/event/{id_event}:
    get:
      tags:
        - events
      summary: Получить мероприятие
      description: Возвращает информацию о мероприятии по ID
      operationId: getEventByID
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о мероприятии
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '404':
          description: Мероприятие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags:
        - events
      summary: Обновить мероприятие
      description: Обновляет информацию о мероприятии
      operationId: updateEvent
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '200':
          description: Мероприятие обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Мероприятие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags:
        - events
      summary: Удалить мероприятие
      description: Удаляет мероприятие по ID
      operationId: deleteEvent
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Мероприятие удалено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Мероприятие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/manage/icons:
    get:
      tags:
        - icons
      summary: Получить список иконок
      description: Возвращает список всех иконок
      operationId: getIcons
      responses:
        '200':
          description: Список иконок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IconDTO'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - icons
      summary: Создать иконку
      description: Создает новую иконку
      operationId: createIcon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IconRequest'
      responses:
        '201':
          description: Иконка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IconDTO'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/manage/icons/{id}:
    get:
      tags:
        - icons
      summary: Получить иконку
      description: Возвращает информацию об иконке по ID
      operationId: getIconByID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация об иконке
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IconDTO'
        '404':
          description: Иконка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags:
        - icons
      summary: Обновить иконку
      description: Обновляет информацию об иконке
      operationId: updateIcon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IconRequest'
      responses:
        '200':
          description: Иконка обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IconDTO'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Иконка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags:
        - icons
      summary: Удалить иконку
      description: Удаляет иконку
      operationId: deleteIcon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Иконка удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Иконка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Авторизация через Bearer токен
  schemas:
    ErrorResponse:
      type: object
      required:
        - id
        - error
      properties:
        id:
          type: string
          description: ID запроса
        error:
          $ref: '#/components/schemas/ErrorResponseDetail'
    ErrorResponseDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Код ошибки
          example: "validation"
        message:
          type: string
          description: Сообщение об ошибке
          example: "invalid input"
        data:
          $ref: '#/components/schemas/ErrorResponseAdditionalData'
    ErrorResponseAdditionalData:
      type: object
      properties:
        slug:
          type: string
          description: Дополнительная информация
          example: "slug"
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Статус успеха операции
    EventRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Название мероприятия
        description:
          type: string
          description: Описание мероприятия
        category_id:
          type: integer
          description: ID категории
        members:
          $ref: '#/components/schemas/EventMembersDTO'
    EventResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID мероприятия
        name:
          type: string
          description: Название мероприятия
        description:
          type: string
          description: Описание мероприятия
        category_id:
          type: integer
          description: ID категории
        photo_id:
          type: string
          description: UUID фото
        balance:
          type: integer
          description: Баланс мероприятия
    EventListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventResponse'
    EventMembersDTO:
      type: object
      properties:
        user_ids:
          type: array
          items:
            type: integer
            format: int64
          description: Список ID пользователей
        dummies_names:
          type: array
          items:
            type: string
          description: Список имен dummy-пользователей
...
    IconRequest:
      type: object
      required:
        - name
        - file_uuid
      properties:
        name:
          type: string
          description: Название иконки
        file_uuid:
          type: string
          description: UUID файла иконки
    IconDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID иконки
        name:
          type: string
          description: Название иконки
        file_uuid:
          type: string
          description: UUID файла иконки
        external_uuid:
          type: string
          description: Внешний UUID
...